pre_total_requirements.txt
absl-py==0.10.0
appdirs==1.4.4
asgiref==3.2.10
astunparse==1.6.3
audioread==2.1.8
autopep8==1.5.4
backcall==0.2.0
beautifulsoup4==4.6.0
cachetools==4.1.1
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
click==7.1.2
colorama==0.4.3
cycler==0.10.0
Cython==0.29.12
decorator==4.4.2
defusedxml==0.6.0
dill==0.3.2
Django==3.1.1
django-allauth==0.42.0
django-appconf==1.0.4
django-cors-headers==3.5.0
django-imagekit==4.0.2
django-rest-auth==0.9.5
djangorestframework==3.11.1
djangorestframework-jwt==1.11.0
future==0.18.2
gast==0.3.3
google-auth==1.21.3
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
grpcio==1.32.0
gunicorn==20.0.4
h5py==2.10.0
idna==2.10
importlib-metadata==2.0.0
inflect==4.1.0
ipython==7.16.1
ipython-genutils==0.2.0
jamo==0.4.1
jedi==0.17.2
joblib==0.16.0
JPype1==1.0.2
Keras-Preprocessing==1.1.2
kiwisolver==1.2.0
konlpy==0.5.2
librosa==0.7.0
llvmlite==0.31.0
lxml==4.5.2
Markdown==3.2.2
matplotlib==3.3.2
matterhook==0.2
mecab-python==1.0.0
mecab-python3==1.0.1
multiprocess==0.70.10
nltk==3.5
numba==0.48.0
numpy==1.18.5
oauthlib==3.1.0
opt-einsum==3.3.0
packaging==20.4
parso==0.7.1
pathos==0.2.6
pexpect==4.8.0
pickleshare==0.7.5
pilkit==2.0
Pillow==7.2.0
pkg-resources==0.0.0
pooch==1.2.0
pox==0.2.8
ppft==1.6.6.2
prompt-toolkit==3.0.7
protobuf==3.13.0
ptyprocess==0.6.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.6.0
pycparser==2.20
Pygments==2.7.1
PyJWT==1.7.1
pyparsing==2.4.7
PySocks==1.7.1
PySoundFile==0.9.0.post1
python-dateutil==2.8.1
python3-openid==3.2.0
pytz==2020.1
pyword==1.0
PyYAML==5.3.1
regex==2020.7.14
requests==2.24.0
requests-oauthlib==1.3.0
resampy==0.2.2
rsa==4.6
scikit-learn==0.23.2
scipy==1.4.1
six==1.15.0
SoundFile==0.10.3.post1
sqlparse==0.3.1
supervisor==4.2.1
tensorboard==2.3.0
tensorboard-plugin-wit==1.7.0
tensorflow-addons==0.11.2
tensorflow-estimator==2.3.0
tensorflow-gpu==2.3.0
termcolor==1.1.0
threadpoolctl==2.1.0
toml==0.10.1
torch==1.5.1
tqdm==4.49.0
traitlets==4.3.3
tweepy==3.9.0
typeguard==2.9.1
typing-extensions==3.7.4.3
Unidecode==1.1.1
urllib3==1.25.10
wcwidth==0.2.5
Werkzeug==1.0.1
wrapt==1.12.1
zipp==3.2.0
